+++ components/camel-weather/src/test/java/org/apache/camel/component/weather/CurrentWeatherConsumerIT.java
@@ -17,10 +17,13 @@
 package org.apache.camel.component.weather;
 
 import org.apache.camel.builder.RouteBuilder;
+import org.junit.jupiter.api.condition.EnabledIfSystemProperty;
 
 import static org.apache.camel.test.junit5.TestSupport.assertStringContains;
 
-public class CurrentWeatherConsumerTest extends BaseWeatherConsumerTest {
+@EnabledIfSystemProperty(named = "enable.weather.tests", matches = "true",
+                         disabledReason = "Disabled to avoid hitting API limits")
+public class CurrentWeatherConsumerIT extends BaseWeatherConsumerIT {
 
     @Override
     protected void checkWeatherContent(String weather) {
