+++ components/camel-weather/src/test/java/org/apache/camel/component/weather/CurrentWeatherMadridProducerIT.java
@@ -21,11 +21,14 @@ import org.apache.camel.Message;
 import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.component.mock.MockEndpoint;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.condition.EnabledIfSystemProperty;
 
 import static org.apache.camel.test.junit5.TestSupport.assertIsInstanceOf;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
-public class CurrentWeatherMadridProducerTest extends BaseWeatherConsumerTest {
+@EnabledIfSystemProperty(named = "enable.weather.tests", matches = "true",
+                         disabledReason = "Disabled to avoid hitting API limits")
+public class CurrentWeatherMadridProducerIT extends BaseWeatherConsumerIT {
 
     @Override
     @Test
